+++ pkg/core/resources/apis/mesh/dataplane_validator_test.go
@@ -90,6 +90,21 @@ var _ = Describe("Dataplane", func() {
                 - port: 3333
                   service: redis`,
 		),
+		Entry("dataplane with valid tags", `
+            type: Dataplane
+            name: dp-1
+            mesh: default
+            networking:
+              address: 192.168.0.1
+              gateway:
+                tags:
+                  service: backend
+                  version: "1"
+                  valid: abc.0123-789.under_score:90
+              outbound:
+                - port: 3333
+                  service: redis`,
+		),
 	)
 
 	type testCase struct {
@@ -554,6 +569,48 @@ var _ = Describe("Dataplane", func() {
                   message: interface cannot be defined with port. Replace it with port and address
                 - field: networking.outbound[0].interface
                   message: interface cannot be defined with address. Replace it with port and address`,
+		}),
+		Entry("networking.inbound: tag name with invalid characters", testCase{
+			dataplane: `
+                type: Dataplane
+                name: dp-1
+                mesh: default
+                networking:
+                  address: 192.168.0.1
+                  inbound:
+                    - port: 1234
+                      tags:
+                        service: backend
+                        version: "v1"
+                        inv@lidT/gN%me: value
+                  outbound:
+                    - port: 3333
+                      service: redis`,
+			expected: `
+                violations:
+                - field: networking.inbound[0].tags["inv@lidT/gN%me"]
+                  message: tag name must consist of alphanumeric characters, dots, dashes and underscores`,
+		}),
+		Entry("networking.inbound: tag value with invalid characters", testCase{
+			dataplane: `
+                type: Dataplane
+                name: dp-1
+                mesh: default
+                networking:
+                  address: 192.168.0.1
+                  inbound:
+                    - port: 1234
+                      tags:
+                        service: backend
+                        version: "v1"
+                        invalidTagValue: inv@lid+t@g
+                  outbound:
+                    - port: 3333
+                      service: redis`,
+			expected: `
+                violations:
+                - field: networking.inbound[0].tags["invalidTagValue"]
+                  message: tag value must consist of alphanumeric characters, dots, dashes and underscores`,
 		}),
 	)